Create new test file for ops formatted queries
